Split backoffice components by domain
